python_backend/requirements.txt
fastapi==0.109.0
sqlalchemy==2.0.25
passlib==1.7.4
python-jose==3.3.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.5
bcrypt==3.2.0